# frontend/backend-api-docs/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dataspace",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/healthz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Healthz",
        "operationId": "healthz_api_v1_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth",
          "Authentication"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth",
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{space_id}/folders": {
      "post": {
        "tags": [
          "upload",
          "File Management"
        ],
        "summary": "Create Folder",
        "description": "创建目录：自动计算 path_cache\n路径规则：/根目录/子目录",
        "operationId": "create_folder_api_v1_spaces__space_id__folders_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{space_id}/files/upload-init": {
      "post": {
        "tags": [
          "upload",
          "File Management"
        ],
        "summary": "Init Upload",
        "description": "初始化上传：生成预签名地址",
        "operationId": "init_upload_api_v1_spaces__space_id__files_upload_init_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Space Id"
            }
          },
          {
            "name": "folder_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Folder Id"
            }
          },
          {
            "name": "filename",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          },
          {
            "name": "size_bytes",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Size Bytes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{space_id}/files/upload-complete": {
      "post": {
        "tags": [
          "upload",
          "File Management"
        ],
        "summary": "Complete Upload",
        "description": "确认上传完成：同步元数据至 files 和 file_versions",
        "operationId": "complete_upload_api_v1_spaces__space_id__files_upload_complete_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          },
          {
            "name": "object_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{space_id}/files/{file_public_id}/view": {
      "get": {
        "tags": [
          "upload",
          "File Management"
        ],
        "summary": "Get File Content",
        "description": "获取文件的临时查看/下载链接",
        "operationId": "get_file_content_api_v1_spaces__space_id__files__file_public_id__view_get",
        "parameters": [
          {
            "name": "file_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Public Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{space_id}/folders/{folder_public_id}/rename": {
      "patch": {
        "tags": [
          "upload",
          "File Management"
        ],
        "summary": "Rename Folder",
        "description": "重命名目录：不仅更新自身，还需递归更新所有子目录的 path_cache",
        "operationId": "rename_folder_api_v1_spaces__space_id__folders__folder_public_id__rename_patch",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Space Id"
            }
          },
          {
            "name": "folder_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Folder Public Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{space_id}/files/{file_public_id}/rename": {
      "patch": {
        "tags": [
          "upload",
          "File Management"
        ],
        "summary": "Rename File",
        "description": "重命名文件",
        "operationId": "rename_file_api_v1_spaces__space_id__files__file_public_id__rename_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Space Id"
            }
          },
          {
            "name": "file_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Public Id"
            }
          },
          {
            "name": "new_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "New Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{space_id}/tree": {
      "get": {
        "tags": [
          "upload",
          "File Management"
        ],
        "summary": "Get Space Tree",
        "description": "获取指定空间的完整目录树\n算法逻辑：\n1. 一次性查询空间下所有 Folder 和 File\n2. 使用内存映射表（Map）将扁平数据组装成递归嵌套结构",
        "operationId": "get_space_tree_api_v1_spaces__space_id__tree_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TreeFolderResponse"
                  },
                  "title": "Response Get Space Tree Api V1 Spaces  Space Id  Tree Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces": {
      "get": {
        "tags": [
          "space",
          "Spaces"
        ],
        "summary": "Get Spaces",
        "description": "获取当前用户拥有的所有空间，支持分页",
        "operationId": "get_spaces_api_v1_spaces_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpaceResponse"
                  },
                  "title": "Response Get Spaces Api V1 Spaces Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "space",
          "Spaces"
        ],
        "summary": "Create Space",
        "description": "创建新空间并确保事务完整性",
        "operationId": "create_space_api_v1_spaces_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpaceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{spaceId}": {
      "delete": {
        "tags": [
          "space",
          "Spaces"
        ],
        "summary": "Delete Space",
        "description": "物理/逻辑删除空间，并检查最小保留约束",
        "operationId": "delete_space_api_v1_spaces__spaceId__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spaceid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{spaceId}/switch": {
      "post": {
        "tags": [
          "space",
          "Spaces"
        ],
        "summary": "Switch Space",
        "description": "切换当前活跃空间。\n在分布式架构中，此处通常需要更新 Redis 中的 'current_space_id' 映射。",
        "operationId": "switch_space_api_v1_spaces__spaceId__switch_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spaceid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthRequest": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "credential": {
            "type": "string",
            "title": "Credential"
          },
          "identity_type": {
            "type": "string",
            "title": "Identity Type",
            "default": "password"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "credential"
        ],
        "title": "AuthRequest"
      },
      "AuthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "user_id",
          "message"
        ],
        "title": "AuthResponse"
      },
      "FileBrief": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "name",
          "size_bytes",
          "mime"
        ],
        "title": "FileBrief"
      },
      "FolderCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "文件夹名称"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "父文件夹ID，根目录传空"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "FolderCreate"
      },
      "FolderRenameRequest": {
        "properties": {
          "new_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "New Name"
          }
        },
        "type": "object",
        "required": [
          "new_name"
        ],
        "title": "FolderRenameRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SpaceCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SpaceCreate"
      },
      "SpaceResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner_user_id": {
            "type": "integer",
            "title": "Owner User Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "public_id",
          "name",
          "owner_user_id"
        ],
        "title": "SpaceResponse"
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token"
      },
      "TreeFolderResponse": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path_cache": {
            "type": "string",
            "title": "Path Cache"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/TreeFolderResponse"
            },
            "type": "array",
            "title": "Children",
            "default": []
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileBrief"
            },
            "type": "array",
            "title": "Files",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "name",
          "path_cache"
        ],
        "title": "TreeFolderResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}